+++ examples/instability/BUILD.bazel
@@ -3,30 +3,39 @@
 
 load(
     "@drake//tools/skylark:drake_py.bzl",
+    "drake_py_library",
     "drake_py_unittest",
 )
 load("//tools/lint:lint.bzl", "add_lint_tests")
 
+drake_py_library(
+    name = "instability_common_py",
+    srcs = ["test/instability_common.py"],
+    deps = ["//bindings/pydrake"],
+)
+
 drake_py_unittest(
     name = "control_instability_test",
-    timeout = "moderate",
+    timeout = "long",
     data = [
         "test/floating_body.sdf",
         "//manipulation/models/franka_description:models",
     ],
     deps = [
+        ":instability_common_py",
         "//bindings/pydrake",
     ],
 )
 
 drake_py_unittest(
     name = "contact_instability_test",
-    timeout = "moderate",
+    timeout = "long",
     data = [
         "test/contact_object_a.sdf",
         "test/contact_object_b.sdf",
     ],
     deps = [
+        ":instability_common_py",
         "//bindings/pydrake",
     ],
 )
