+++ examples/instability/test/control_instability_test.py
@@ -1,8 +1,9 @@
 """
-Shows simulated controllers going unstable.
+Shows simulated controllers going unstable. These are purposeful "over stiff"
+controllers. Concretely, these controllers do not attempt to cancel out any
+dynamics.
 """
 
-from contextlib import contextmanager
 import dataclasses as dc
 from types import SimpleNamespace
 import unittest
@@ -12,63 +13,22 @@ import numpy as np
 from pydrake.common import FindResourceOrThrow
 from pydrake.common.cpp_param import List
 from pydrake.common.value import Value
-from pydrake.geometry import DrakeVisualizer, DrakeVisualizerParams, Role
 from pydrake.multibody.math import SpatialForce, SpatialVelocity
-from pydrake.multibody.parsing import Parser
-from pydrake.multibody.plant import (
-    AddMultibodyPlant,
-    ExternallyAppliedSpatialForce,
-    MultibodyPlantConfig,
-)
-from pydrake.multibody.tree import JacobianWrtVariable
+from pydrake.multibody.plant import ExternallyAppliedSpatialForce
 from pydrake.systems.analysis import Simulator
-from pydrake.systems.framework import DiagramBuilder, EventStatus, LeafSystem
+from pydrake.systems.framework import LeafSystem
+
+from drake.examples.instability.test.instability_common import (
+    TestBase,
+    add_basic_simulation_components,
+    get_frame_spatial_velocity,
+    lstsq,
+    make_force_for_frame,
+    monitor_large_energy_delta,
+    simple_jacobian,
+)
 
 VISUALIZE = False
-
-
-def simple_jacobian(plant, context, frame_W, frame_F):
-    Jv_WF = plant.CalcJacobianSpatialVelocity(
-        context,
-        with_respect_to=JacobianWrtVariable.kV,
-        frame_B=frame_F,
-        p_BP=[0, 0, 0],
-        frame_A=frame_W,
-        frame_E=frame_W,
-    )
-    return Jv_WF
-
-
-def get_frame_spatial_velocity(plant, context, frame_W, frame_F):
-    Jv_WF = simple_jacobian(plant, context, frame_W, frame_F)
-    v = plant.GetVelocities(context)
-    V_WF = SpatialVelocity(Jv_WF @ v)
-    return V_WF
-
-
-def make_force_for_frame(frame_F, F_F_W):
-    external_force = ExternallyAppliedSpatialForce()
-    external_force.body_index = frame_F.body().index()
-    external_force.F_Bq_W = F_F_W
-    external_force.p_BoBq_B = frame_F.GetFixedPoseInBodyFrame().translation()
-    return external_force
-
-
-def add_basic_simulation_components(ns, time_step, solver=None):
-    ns.builder = DiagramBuilder()
-    config = MultibodyPlantConfig(time_step=time_step)
-    if solver is not None:
-        config.discrete_contact_solver = solver
-    else:
-        assert time_step == 0.0
-    ns.plant, ns.scene_graph = AddMultibodyPlant(config, ns.builder)
-    ns.parser = Parser(ns.plant)
-    DrakeVisualizer.AddToBuilder(
-        ns.builder,
-        ns.scene_graph,
-        params=DrakeVisualizerParams(role=Role.kIllustration),
-    )
-    ns.plant.mutable_gravity_field().set_gravity_vector([0.0, 0.0, 0.0])
 
 
 def add_floating_sim(ns):
@@ -114,6 +74,8 @@ class RotationalVelocityDamper:
 
 
 class FloatingController(LeafSystem):
+    """Applies RotationalVelocityDamper to a floating body."""
+
     def __init__(self, plant, frame_F, damper):
         super().__init__()
         frame_W = plant.world_frame()
@@ -153,6 +115,16 @@ class FloatingController(LeafSystem):
 
 
 class ArticulatedController(LeafSystem):
+    """
+    Applies RotationalVelocityDamper to an articulated model instance.
+
+    Note: This may be expected to dampen energy; however, since we are
+    injecting energy to remove rotation velocity, that energy may go towards
+    translation.
+
+    TODO(eric.cousineau): Add translation damping.
+    """
+
     def __init__(self, plant, frame_F, damper):
         super().__init__()
         frame_W = plant.world_frame()
@@ -192,53 +164,7 @@ class ArticulatedController(LeafSystem):
         )
 
 
-def lstsq(A, b):
-    return np.linalg.lstsq(A, b, rcond=None)[0]
-
-
-def norm(x):
-    return np.linalg.norm(x)
-
-
-def total_energy(plant, context):
-    return plant.EvalKineticEnergy(context) + plant.EvalPotentialEnergy(
-        context
-    )
-
-
-def monitor_large_energy_delta(simulator, t, plant, max_energy_gain):
-    diagram_context = simulator.get_context()
-    context = plant.GetMyContextFromRoot(diagram_context)
-    energy_init = total_energy(plant, context)
-    max_bad_energy_delta = None
-
-    def monitor(diagram_context):
-        nonlocal max_bad_energy_delta
-        context = plant.GetMyContextFromRoot(diagram_context)
-        energy_now = total_energy(plant, context)
-        energy_delta = energy_now - energy_init
-        if energy_delta > max_energy_gain:
-            if (
-                max_bad_energy_delta is None
-                or energy_delta > max_bad_energy_delta
-            ):
-                max_bad_energy_delta = energy_delta
-        return EventStatus.DidNothing()
-
-    simulator.set_monitor(monitor)
-    simulator.AdvanceTo(t)
-    if max_bad_energy_delta is not None:
-        raise RuntimeError(f"Too much energy gained: {max_bad_energy_delta} J")
-
-
-class Test(unittest.TestCase):
-    @contextmanager
-    def assert_raises_message(self, pieces, cls=RuntimeError):
-        with self.assertRaises(cls) as cm:
-            yield
-        for piece in pieces:
-            self.assertIn(piece, str(cm.exception))
-
+class Test(TestBase):
     def run_floating(
         self, time_step, *, kd, solver=None, max_energy_gain,
     ):
